+++ kof-compiler/src/main/java/dev/kof/compiler/nat/NativeRiscvAsm.java
@@ -298,6 +298,7 @@ public final class NativeRiscvAsm {
                 .append(NativeRiscvAsmIoMove.RISCV_RUNTIME_ASM_IO_MOVE)
                 .append(NativeRiscvAsmIoCopy.RISCV_RUNTIME_ASM_IO_COPY)
                 .append(NativeRiscvAsmProcess.RISCV_RUNTIME_ASM_PROCESS)
+                .append(NativeRiscvAsmProcessResult.RISCV_RUNTIME_ASM_PROCESS_RESULT)
                 .append(NativeRiscvAsmShell.RISCV_RUNTIME_ASM_SHELL)
                 .append(NativeRiscvAsmPipeline.RISCV_RUNTIME_ASM_PIPELINE)
                 .append(NativeRiscvAsmSsh.RISCV_RUNTIME_ASM_SSH)